requirements.txt
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==5.1.1
Django==5.0.1
exceptiongroup==1.2.0
executing==2.0.1
idna==3.6
ipython==8.20.0
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
postgres==4.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
psycopg2-pool==1.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.13
whitenoise==6.6.0
convert_numbers
django_rest_framework
pybreaker==1.1.0
retry==0.9.2
